Fix after pre-commit
